@@ -1,22 +1,34 @@
-# -*- encoding: utf-8 -*-
-
+# encoding: UTF-8
 Gem::Specification.new do |s|
   s.platform    = Gem::Platform::RUBY
-
   s.name        = 'spree_affiliate'
-  s.version     = '1.0.1'
+  s.version     = '2.0.3'
+ 
   s.authors     = ['Rails Dog']
   s.email       = 'gems@railsdog.com'
   s.homepage    = 'http://github.com/spree/spree_affiliate'
   s.summary     = 'Affiliate support for Spree'
   s.description = 'Affiliate support for Spree'
-  s.required_ruby_version = '>= 1.8.7'
-  s.rubygems_version      = '1.3.6'
 
-  s.files         = `git ls-files`.split("\n")
-  s.test_files    = `git ls-files -- {test,spec,features}/*`.split("\n")
-  s.require_paths = ["lib"]
 
-  s.add_dependency('spree_core', '>= 1.3')
-  s.add_dependency('spree_store_credits', '>= 1.0.0')
+ s.required_ruby_version = '>= 1.9.3'
+
+ 
+  #s.files       = `git ls-files`.split("\n")
+  #s.test_files  = `git ls-files -- {test,spec,features}/*`.split("\n")
+  s.require_path = 'lib'
+  s.requirements << 'none'
+
+  s.add_dependency 'spree_core', '~> 2.0.3'
+
+  s.add_development_dependency 'capybara', '~> 2.1'
+  s.add_development_dependency 'coffee-rails'
+  s.add_development_dependency 'database_cleaner'
+  s.add_development_dependency 'factory_girl', '~> 4.2'
+  s.add_development_dependency 'ffaker'
+  s.add_development_dependency 'rspec-rails',  '~> 2.13'
+  s.add_development_dependency 'sass-rails'
+  s.add_development_dependency 'selenium-webdriver'
+  s.add_development_dependency 'simplecov'
+  s.add_development_dependency 'sqlite3'
 end
